LINUX/requirements.txt
bs4
pillow
--upgrade oauth2client 
httplib2
requests
command
